+++ src/main/java/com/bitbank/exceptions/ControllerAdvisor.java
@@ -3,6 +3,7 @@ package com.bitbank.exceptions;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.AccessDeniedException;
+import org.springframework.security.authentication.BadCredentialsException;
 import org.springframework.web.bind.annotation.ControllerAdvice;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.servlet.mvc.method.annotation.ResponseEntityExceptionHandler;
@@ -59,6 +60,19 @@ public class ControllerAdvisor extends ResponseEntityExceptionHandler {
 	 */
 	@ExceptionHandler(InvalidMfaException.class)
 	protected ResponseEntity<ErrorResponse> handleInvalidMfaException(InvalidMfaException ex) {
+		return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(new ErrorResponse(ex.getMessage()));
+	}
+	
+	/**
+	 * Handle BadCredentialsException
+	 * 
+	 * Return the entity with Error response.
+	 * 
+	 * @param ex
+	 * @return
+	 */
+	@ExceptionHandler(BadCredentialsException.class)
+	protected ResponseEntity<ErrorResponse> handleBadCredentialsException(BadCredentialsException ex) {
 		return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(new ErrorResponse(ex.getMessage()));
 	}